requirements.txt
WTForms==2.3.3
Flask==2.0.2
Flask_WTF==0.15.1
Flask_SQLAlchemy==2.5.1
PyJWT==1.7.0

flaskweb~=0.2.1
bs4~=0.0.1
beautifulsoup4~=4.10.0
selenium~=4.0.0